Removes chromedriver version from pip install
This is in an attempt to standardize the build process, in order to
migrate it to the new recommended approach of using the containerized
builds in travis.

The idea is to make the travis build system a python one, so every
single python dependency should be in the requirements.txt file.

The containerized builds do not allow for sudo commands anywhere on the
build process, so any sudo-based commands should be run through the
standard build steps, or through some addon. Apt has a travis addon
which allows installing software, but there is no addon for pip modules.
The only way to install pip modules is through the python container.
This commit makes every python dependency be in requirements.txt, so
that we can later on migrate to a standard python container instead.