Makes travis use the containerized infrastructure
Makes the project a python project so that travis automatically installs
all pip dependencies in the requirements.txt file.

Moves the python and node version to the standard travis sections, so
that they are managed by travis itself.

Moves apt packages out of the prerequisites step and into the standard
travis apt addon.