Adds logic to manually release sources once animations are destroyed
We have an issue where sources are added on selection (when a selection
animation is created) but they are never released (when destroying the
selection animation, for instance). Because of this, the sources begin
accumulating on each selection and increase the number of requests when
zooming in or moving the map, as all tiled sources need to be udpated
with the new bounding box.

This commit adds a destroyView method which releases all resources
created when a DataView is created inside the DataManager. Currently,
the only resources that are released are the data sources that are
registered when creating the DataView.