Fixes invalid [Object object] legend
The selection event data field has a toString method, which surprisingly
doesn't return an actual string, but instead returns a jquery node. This
allows the originator of the event data to attach event handlers to the
node, which is then appended into the info window.

We are not refactoring that method into something else in this branch
though. I created a ticket at #109.