Implements multivalue parameter parsing in reports
This commit allows properties from selected cartodb polygons to export
values for interpolation in the layer's urlTemplate and promptTemplate
report properties. Values may either be direct values, meaning that the
value is directly available in the interpolation context, or
multivalued values, meaning that the polygon value is instead a string
containing multiple options, and that the user will have to select one
of them to actually generate the report.