Adds info buttons to layer list
The simple layer list used in the normal user sidebar now includes a
small link to a page with additional information about the layer.

The url for that is stored in the workspace for each animation. This has
the problem that workspaces will need to be updated when a value changes
for these kinds of things, but it's the best we can do right now as we
have already discussed we need somewhere to put layer metadata and the
changes needed for that exceed this ticket.